Types

State.NLPAtX — Type

NLPAtX contains the important information concerning a non linear problem at the iteration x. Basic information is:

x the current candidate for solution to our original problem

f(x) which is the funciton evaluation at x

g(x) which is the gradient evaluation at x

Hx which is the hessian representation at x

mu : Lagrange multiplier of the bounds constraints

cx : evaluation of the constraint function at x

Jx : jacobian matrix of the constraint function at x

lambda : Lagrange multiplier of the constraints

start_time : Default is a NaN, can be updated to fit the start of the algorithm.

evals : number of evaluations of the function (import the type NLPModels.Counters)

All these information (except for x) are optionnal and need to be update when required. The update is done trhough the update! function.

State.LSAtT — Type

A structure designed to track line search information from one iteration to another. If we have f : ℜⁿ → ℜ, then we define h(θ) = f(x + θ*d) where x and d are vectors of same dimension and θ is a scalar, more specifically our step size.

Tracked data can include:

x : our current step size

ht : h(θ) at the current iteration

gt : h'(θ)

h₀ : h(0)

g₀ : h'(0)

start_time: the time at which the line search algorithm started.

Unless they are defined otherwise, the default value for all parameter is NaN (except for x). They can be updated through the update! function.

Example:

ls_a_t = LSAtT(1.0)
update!(ls_a_t, x = 0.0, h₀ = obj(h, 0.0), g₀ = grad(h, 0.0))

Function

State.update! — Function

Updates the (desired) values of an object of type NLPAtX. Inputs:

An NLPAtX object

Any keywords that needs to be updated.